packages/tether-python/pyproject.toml
[project]
name = "tether"
version = "0.1.0"
description = "Python backend utilities for Tether AI/ML desktop applications"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Tether Contributors" }
]
keywords = ["fastapi", "ai", "ml", "llm", "tauri"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
openai = [
    "openai>=1.0.0",
]
local = [
    "llama-cpp-python>=0.2.0",
]
all = [
    "openai>=1.0.0",
    "llama-cpp-python>=0.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[project.urls]
Homepage = "https://github.com/your-org/tether"
Documentation = "https://github.com/your-org/tether#readme"
Repository = "https://github.com/your-org/tether"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tether"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
